Move skip-gram code to word_embedding dir
